+++ packages/react/lib/DisplayableSettings/Property.tsx
@@ -10,10 +10,11 @@ import {
 import { get, classNames } from '@junipero/react';
 
 import Text from '../Text';
+import { ReactComponentSettingsFieldObject } from '../types';
 
 interface PropertyProps extends ComponentPropsWithoutRef<'span'> {
   element: ElementObject;
-  field: ComponentSettingsFieldObject;
+  field: ReactComponentSettingsFieldObject;
   override?: ComponentOverrideObject | ComponentOverride;
 }
 
@@ -28,7 +29,6 @@ const Property = ({
     setting,
     override?.fields?.find(f => f.key === setting.key) || {},
   ), [setting, override]);
-
   const value = useMemo(() => (
     get(element, field.key as string, typeof field.default === 'function'
       ? field.default(element) : field.default)
@@ -50,7 +50,7 @@ const Property = ({
         { option?.title ? (
           <Text>{ option.title }</Text>
         ) : (
-          <Text>{ value }</Text>
+          <Text>{ setting.display(value) }</Text>
         ) }
       </span>
     </span>