+++ packages/addon-ckeditor5-react/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.12.5](https://github.com/p3ol/oak/compare/v3.12.4...v3.12.5) (2025-06-17)
+
+**Note:** Version bump only for package @oakjs/addon-ckeditor5-react
+
+
+
+
+
+## [3.12.4](https://github.com/p3ol/oak/compare/v3.12.3...v3.12.4) (2025-06-11)
+
+**Note:** Version bump only for package @oakjs/addon-ckeditor5-react
+
+
+
+
+
+## [3.12.3](https://github.com/p3ol/oak/compare/v3.12.2...v3.12.3) (2025-06-11)
+
+**Note:** Version bump only for package @oakjs/addon-ckeditor5-react
+
+
+
+
+
+## [3.12.2](https://github.com/p3ol/oak/compare/v3.12.1...v3.12.2) (2025-04-15)
+
+**Note:** Version bump only for package @oakjs/addon-ckeditor5-react
+
+
+
+
+
+## [3.12.1](https://github.com/p3ol/oak/compare/v3.12.0...v3.12.1) (2025-03-13)
+
+**Note:** Version bump only for package @oakjs/addon-ckeditor5-react
+
+
+
+
+
 # [3.12.0](https://github.com/p3ol/oak/compare/v3.11.2...v3.12.0) (2025-02-03)
 
 **Note:** Version bump only for package @oakjs/addon-ckeditor5-react